fix(websocket): remove token manager and no websocket auth token re-use
